Better document PWA stuff
