+++ packages/react-scripts/template/src/index.js
@@ -1,9 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import registerServiceWorker from './service-worker-registration';
+import registerServiceWorker from './registerServiceWorker';
 import './index.css';
 
 ReactDOM.render(<App />, document.getElementById('root'));
-
 registerServiceWorker();
